+++ libs/ANS/Cache/Interfaces/Memcached.php
@@ -6,6 +6,7 @@ class Memcached implements \ANS\Cache\Icache
     public $loaded = false;
 
     private $settings = array();
+    private $reload = false;
 
     /**
      * public function __construct ([array $settings])
@@ -64,6 +65,10 @@ class Memcached implements \ANS\Cache\Icache
     */
     public function exists ($key)
     {
+        if ($this->reload) {
+            return false;
+        }
+
         if ($this->server->add($key, null)) {
             $this->server->delete($key);
 
@@ -133,5 +138,17 @@ class Memcached implements \ANS\Cache\Icache
     public function expire ($key)
     {
         return false;
+    }
+
+    /**
+    * public function reload (void)
+    *
+    * Allow to skip a cache read and store it again
+    *
+    * return mixed
+    */
+    public function reload ()
+    {
+        $this->reload = true;
     }
 }
